TP2/pyproject.toml
[project]
name = "tp2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.2",
    "pinecone>=7.3.0",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "tqdm>=4.67.3",
]
